6558: Bringing back agent changes from hirt's GitHub fork
Contributed-by: Jessye Coleman-Shapiro<jescolem@redhat.com>, jmatsuoka, hirt

+++ core/org.openjdk.jmc.agent/src/main/java/org/openjdk/jmc/agent/converters/DoubleConverter.java
@@ -43,5 +43,5 @@ public interface DoubleConverter<T> {
 	 *            the object to convert.
 	 * @return the object converted to a double.
 	 */
-	long convert(T o);
+	double convert(T o);
 }
